♻️ [core] improve core maintainability
